refactor: (admin/moderation) extraire helper function à un autre fichier
